Add files via upload
This is a Table Tennis automation product which helps in notifying the occupancy of the TT court by providing information on no of balls and bats available inside court by using the Internet Of Things (ADAFRUIT). So by this, people in the office will be able to know the right time for refreshment (TT) and can plan to use the court accordingly to the occupancy of place. By this, I would provide the value of reducing the wastage of time by unwanted trips of people to the court(TT).